requirements.txt
litellm>=1.83.9
ollama>=0.6.1
python-dotenv
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
websockets>=13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
